\name{circRNA_classb}
\alias{circRNA_classb}
\docType{data}
\title{
circRNA classification data
}
\description{
This data is generated by r package FcircSEC.
}
\usage{data("circRNA_classb")}
\format{
  A data frame with 14 observations on the following 3 variables.
  \describe{
    \item{\code{V1}}{Chromosome names}
    \item{\code{V2}}{Start position }
    \item{\code{V3}}{End position}
  }
}
\details{
This is the output generated by the circClassification function of the r package FcircSEC.
It has three columns: chromosome name, start position and end position of circRNAs.
}
\source{
This example data is produced by r package FcircSEC.
}
\references{
R package FcircSEC.
}
\examples{
data(circRNA_classb)
## maybe str(circRNA_classb) ; plot(circRNA_classb) ...
}
\keyword{datasets}
