\name{circRNA_classt}
\alias{circRNA_classt}
\docType{data}
\title{
circRNA classification data
}
\description{
This data is generated by r package FcircSEC.
}
\usage{data("circRNA_classt")}
\format{
  A data frame with 14 observations on the following 15 variables.
  \describe{
    \item{\code{ID}}{Circular RNA ID}
	
	\item{\code{chr}}{Chromosome}
    \item{\code{circ_start}}{Start position}
    \item{\code{circ_end}}{End position}
    \item{\code{circ_strand}}{Strand}
    \item{\code{splice_L}}{circRNA length}
    \item{\code{circ_type}}{circRNA type}
    \item{\code{e_count}}{Number of exons}
    \item{\code{e_sizes}}{Exon sizes}
	\item{\code{e_offsets}}{Start positions of exons}
	\item{\code{b_transcript}}{Transcript name}
	\item{\code{b_strand}}{Transcript strand}
    \item{\code{b_trans_start}}{Transcript start}
    \item{\code{b_trans_end}}{Transcript end}
    \item{\code{b_gene}}{Host gene}
  }
}
\details{
This is the output generated by the circClassification function of the r package FcircSEC.
It contains all necessary information for extracting the full length circRNA sequence.
}
\source{
This example data is produced by r package FcircSEC.
}
\references{
R package FcircSEC.
}
\examples{
data(circRNA_classt)
## maybe str(circRNA_classt) ; plot(circRNA_classt) ...
}
\keyword{datasets}
