\name{circRNA_genomic_sequence}
\alias{circRNA_genomic_sequence}
\docType{data}
\title{
Genomic sequences of circRNAs.
}
\description{
This data is generated by r package FcircSEC.
}
\usage{data("circRNA_genomic_sequence")}
\format{
  A data frame with 14 observations on the following 2 variables.
  \describe{
    \item{\code{seq_name1}}{IDs of circular RNAs}
	\item{\code{sequence1}}{Genomic sequences of circular RNAs}
  }
}
\details{
This is the output generated by the get.fasta function of the r package FcircSEC.
It contains the genomic sequences of the circRNAs.
}
\source{
This example data is produced by r package FcircSEC.
}
\references{
R package FcircSEC.
}
\examples{
data(circRNA_genomic_sequence)
## maybe str(circRNA_genomic_sequence) ; plot(circRNA_genomic_sequence) ...
}
\keyword{datasets}
